+++ DotDoc.EntityFrameworkCore.Extensions.Tests/ExecuteUpdate/ExecuteInsertTests.cs
@@ -1,0 +1,94 @@
+﻿// Copyright ©2021-2025 Mike King.
+// This file is licensed to you under the MIT license.
+// See the License.txt file in the solution root for more information.
+
+using DotDoc.EntityFrameworkCore.Extensions.DatabaseType;
+using DotDoc.EntityFrameworkCore.Extensions.ExecuteUpdate;
+using DotDoc.EntityFrameworkCore.Extensions.Tests.Data;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Query;
+
+namespace DotDoc.EntityFrameworkCore.Extensions.Tests.ExecuteUpdate;
+
+/// <summary>
+/// Tests for ExecuteInsert extensions.
+/// </summary>
+[TestClass]
+public class ExecuteInsertTests
+{
+    #region public methods
+
+    /// <summary>
+    /// Test ExecuteInsertAsync with Null IQueryable source parameter.
+    /// </summary>
+    /// <returns>A <see cref="Task"/> that represents the asynchronous operation.</returns>
+    [TestMethod(DisplayName = "ExecuteInsertAsync with Null IQueryable source parameter")]
+    public async Task ExecuteInsertTests_001_Async()
+    {
+        // ARRANGE
+        IQueryable<TestTable1> source = null!;
+        Action<UpdateSettersBuilder<TestTable1>> setPropertyCalls = new(builder => { });
+
+        // ACT / ASSERT
+        await Assert.ThrowsExactlyAsync<ArgumentNullException>(() => source.ExecuteInsertAsync(setPropertyCalls, CancellationToken.None), "Unexpected exception").ConfigureAwait(false);
+    }
+
+    /// <summary>
+    /// Test ExecuteInsertAsync with Null UpdateSettersBuilder setPropertyCalls parameter.
+    /// </summary>
+    /// <returns>A <see cref="Task"/> that represents the asynchronous operation.</returns>
+    [TestMethod(DisplayName = "ExecuteInsertAsync with Null UpdateSettersBuilder setPropertyCalls parameter")]
+    public async Task ExecuteInsertTests_002_Async()
+    {
+        // ARRANGE
+        IQueryable<TestTable1> source = Array.Empty<TestTable1>().AsQueryable();
+        Action<UpdateSettersBuilder<TestTable1>> setPropertyCalls = null!;
+
+        // ACT / ASSERT
+        await Assert.ThrowsExactlyAsync<ArgumentNullException>(() => source.ExecuteInsertAsync(setPropertyCalls, CancellationToken.None), "Unexpected exception").ConfigureAwait(false);
+    }
+
+    /// <summary>
+    /// Test ExecuteInsertAsync.
+    /// </summary>
+    /// <param name="databaseType">Database type.</param>
+    /// <returns>A <see cref="Task"/> that represents the asynchronous operation.</returns>
+    [TestMethod(DisplayName = "ExecuteInsertAsync")]
+    [DataRow(DatabaseTypes.Sqlite, DisplayName = DatabaseTypes.Sqlite)]
+    [DataRow(DatabaseTypes.SqlServer, DisplayName = DatabaseTypes.SqlServer)]
+    public async Task ExecuteInsertTests_003_Async(string databaseType)
+    {
+        // ARRANGE
+        using Context context = await DatabaseUtils.CreateDatabaseAsync(
+            databaseType: databaseType,
+            customConfigurationActions: (optionsBuilder) =>
+            {
+                optionsBuilder.UseExecuteUpdateExtensions();
+            })
+            .ConfigureAwait(false);
+
+        long id = 1;
+        string value = "TestValue";
+
+        // ACT
+        await context.TestTable1.ExecuteInsertAsync(
+            builder =>
+            {
+                builder.SetProperty(e => e.TestField, value);
+            },
+            CancellationToken.None)
+            .ConfigureAwait(false);
+
+        // ASSERT
+        TestTable1? row = await context.TestTable1
+            .Where(e => e.Id == id)
+            .SingleOrDefaultAsync(CancellationToken.None)
+            .ConfigureAwait(false);
+
+        Assert.IsNotNull(row, "No row returned");
+        Assert.AreEqual(id, row.Id, "Unexpected Id value");
+        Assert.AreEqual(value, row.TestField, "Unexpected field value");
+    }
+
+    #endregion public methods
+}